Check User role

+++ src/main/java/org/example/e_commerce/Service/CategoryService.java
@@ -3,6 +3,7 @@ package org.example.e_commerce.Service;
 import org.example.e_commerce.Entity.Category;
 import org.example.e_commerce.Repository.CategoryRepository;
 import org.example.e_commerce.dto.dtoResponse.CategoryResponseDTO;
+import org.example.e_commerce.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +16,7 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepo;
 
+    JwtUtil jwtUtil = new JwtUtil();
     public List<Category> getAllCategories() {
         return categoryRepo.findAll();
     }
@@ -28,12 +30,21 @@ public class CategoryService {
         }
     }
 
-    public CategoryResponseDTO createCategory(Category category) {
+    public CategoryResponseDTO createCategory(Category category , String token) {
+        String role = jwtUtil.extractRole(token);
         Category savedCategory = categoryRepo.save(category);
+        if (role.equals("USER")){
+            return new CategoryResponseDTO("You do not have the necessary permissions to perform this action.", 403L, null);
+        }
         return new CategoryResponseDTO("Category created successfully", 201L, savedCategory);
     }
 
-    public CategoryResponseDTO updateCategory(Long id, Category category) {
+    public CategoryResponseDTO updateCategory(Long id, Category category , String token) {
+
+        String role = jwtUtil.extractRole(token);
+        if (role.equals("USER")){
+            return new CategoryResponseDTO("You do not have the necessary permissions to perform this action.", 403L, null);
+        }
         if (categoryRepo.existsById(id)) {
             category.setCategoryid(id);
             Category updatedCategory = categoryRepo.save(category);
@@ -43,7 +54,11 @@ public class CategoryService {
         }
     }
 
-    public CategoryResponseDTO deleteCategory(Long id) {
+    public CategoryResponseDTO deleteCategory(Long id , String token) {
+        String role = jwtUtil.extractRole(token);
+        if (role.equals("USER")){
+            return new CategoryResponseDTO("You do not have the necessary permissions to perform this action.", 403L, null);
+        }
         if (categoryRepo.existsById(id)) {
             categoryRepo.deleteById(id);
             return new CategoryResponseDTO("Category deleted successfully", 200L, null);
